requirements.txt
fastapi==0.104.1
httpx==0.25.0
PyYAML==6.0.1
uvicorn==0.23.2
